Amended group size so that NSAs are only removed if the NSA bird has >1 NSA statuses within a field period.
Keep birds that have one NSA status within a field period as this indicates they were seen in the territory the season before and after the NSA status. We can be fairly certain that this bird was resident in the NSA territory and wasn't seen.

Birds with >1 NSA status are usually seen in one NSA territory the season before the NSA status, and in the different NSA territory the season after the NSA status. Therefore, the bird could have been resident in either of the NSA territories. There is not reliable way to automate choosing the NSA territory the focal bird was most likely to be resident in. Therefore, remove any individual with >1 status group group size lists.